64 bit fixes.
git-svn-id: https://cvet.by:8443/svn/FOnlineSource/Main@1106 c2c6a35e-ad2b-4e1f-bb64-97b989be360b

Former-commit-id: e494d2229bb8a7abde301457a2f15358cdb08cd5

+++ Source/Mapper.cpp
@@ -4789,7 +4789,7 @@ void FOMapper::SScriptFunc::Global_DeleteTile( ushort hx, ushort hy, bool roof, 
             if( tiles[ i ].Layer == layer )
             {
                 tiles.erase( tiles.begin() + i );
-                f.EraseTile( i, roof );
+                f.EraseTile( (uint) i, roof );
                 deleted = true;
                 break;
             }
